requirements.txt
streamlit
google-generativeai
easyocr
